update function signature

+++ src/esp_dmx.h
@@ -294,7 +294,7 @@ esp_err_t dmx_read_packet(dmx_port_t dmx_num, uint8_t *buffer, uint16_t size);
  * @retval ESP_ERR_INVALID_ARG if there was an argument error.
  * @retval ESP_ERR_INVALID_STATE if the driver was not installed.
  */
-esp_err_t dmx_read_slot(dmx_port_t dmx_num, int slot_idx, uint8_t *value);
+esp_err_t dmx_read_slot(dmx_port_t dmx_num, uint16_t slot_idx, uint8_t *value);
 
 /**
  * @brief Write data to the DMX driver from a given buffer and length.
@@ -321,7 +321,7 @@ esp_err_t dmx_write_packet(dmx_port_t dmx_num, const uint8_t *buffer, uint16_t s
  * @retval ESP_ERR_INVALID_ARG if there was an argument error.
  * @retval ESP_ERR_INVALID_STATE if the driver was not installed.
  */
-esp_err_t dmx_write_slot(dmx_port_t dmx_num, int slot_idx, uint8_t value);
+esp_err_t dmx_write_slot(dmx_port_t dmx_num, uint16_t slot_idx, uint8_t value);
 
 /**
  * @brief Transmits a packet of DMX. This sends the number of slots as was
